REFAC: Rename test_osm.py and add import_from_open_street_map test with JSON verification
